#CSW-000: Shivani| Chauhan : Remove unwanted input options provided while creating project from template
