+++ src/main/g8/$assemblyComponentName__space,norm$/src/main/resources/reference.conf
