+++ src/main/g8/$assemblyComponentName__space,norm$/src/main/scala/$organization__lower,packaged$/$subsystem__lower$/$assemblyComponentName__space,norm,word$/$assemblyComponentName__space,Camel$BehaviorFactory.scala
@@ -1,0 +1,13 @@
+package $organization$.$subsystem;format="lower"$.$assemblyComponentName;format="space,norm,word"$;
+import akka.actor.typed.scaladsl.ActorContext
+import csw.framework.models.CswContext
+import csw.framework.scaladsl.{ComponentBehaviorFactory, ComponentHandlers}
+import csw.command.client.messages.TopLevelActorMessage
+
+
+class $assemblyComponentName;format="space,Camel"$BehaviorFactory extends ComponentBehaviorFactory {
+
+  override def handlers(ctx: ActorContext[TopLevelActorMessage], cswCtx: CswContext): ComponentHandlers =
+    new $assemblyComponentName;format="space,Camel"$Handlers(ctx, cswCtx)
+
+}
