+++ src/main/g8/$assemblyComponentName__space,norm$/src/test/resources/application.conf
