+++ src/main/g8/$assemblyComponentName__space,norm$/src/test/scala/$organization__lower,packaged$/$subsystem__lower$/$assemblyComponentName__space,norm,word$/$assemblyComponentName__space,Camel$Test.scala
@@ -1,0 +1,29 @@
+package $organization$.$subsystem;format="lower"$.$assemblyComponentName;format="space,norm,word"$;
+
+import csw.location.api.models.Connection.AkkaConnection
+import csw.prefix.models.Prefix
+import csw.location.api.models.{ComponentId, ComponentType}
+import csw.testkit.scaladsl.CSWService.{AlarmServer, EventServer}
+import csw.testkit.scaladsl.ScalaTestFrameworkTestKit
+import org.scalatest.funsuite.AnyFunSuiteLike
+
+import scala.concurrent.Await
+import scala.concurrent.duration._
+
+class $assemblyComponentName;format="space,Camel"$Test extends ScalaTestFrameworkTestKit(AlarmServer, EventServer) with AnyFunSuiteLike {
+
+  import frameworkTestKit.frameworkWiring._
+
+  override def beforeAll(): Unit = {
+    super.beforeAll()
+    // uncomment if you want one Assembly run for all tests
+    spawnStandalone(com.typesafe.config.ConfigFactory.load("$assemblyComponentName;format="space,Camel"$Standalone.conf"))
+  }
+
+  test("Assembly should be locatable using Location Service") {
+    val connection = AkkaConnection(ComponentId(Prefix("$subsystem$.$assemblyComponentName;format="package"$"), ComponentType.Assembly))
+    val akkaLocation = Await.result(locationService.resolve(connection, 10.seconds), 10.seconds).get
+
+    akkaLocation.connection shouldBe connection
+  }
+}
