--- src/main/g8/$assembly_module$/src/test/resources/$assemblyComponentName__space,Camel$Standalone.conf
@@ -1,4 +1,0 @@
-prefix = "$subsystem$.$assemblyComponentName;format="package"$"
-componentType = assembly
-behaviorFactoryClassName = "org.tmt.$subsystem;format="lower"$.$assemblyComponentName;format="space,norm,word"$;.$assemblyComponentName;format="space,Camel"$BehaviorFactory"
-locationServiceUsage = RegisterOnly
