--- src/main/g8/$deploy_module$/src/main/resources/$hcdComponentName__space,Camel$Standalone.conf
@@ -1,4 +1,0 @@
-prefix = "$subsystem$.$hcdComponentName;format="package"$"
-componentType = hcd
-behaviorFactoryClassName = "org.tmt.$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$;.$hcdComponentName;format="space,Camel"$BehaviorFactory"
-locationServiceUsage = RegisterOnly
