--- src/main/g8/$deploy_module$/src/main/resources/J$hcdComponentName__space,Camel$Standalone.conf
@@ -1,4 +1,0 @@
-prefix = "$subsystem$.$hcdComponentName;format="package"$"
-componentType = hcd
-behaviorFactoryClassName = "org.tmt.$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$;.J$hcdComponentName;format="space,Camel"$BehaviorFactory"
-locationServiceUsage = RegisterOnly
