--- src/main/g8/$deploy_module$/src/main/scala/org/tmt/$subsystem__lower$/$deploy_module__word$/$name__space,Camel$ContainerCmdApp.scala
@@ -1,10 +1,0 @@
-package org.tmt.$subsystem;format="lower"$.$name;format="space,norm,word"$deploy;
-
-import csw.framework.deploy.containercmd.ContainerCmd
-import csw.prefix.models.Subsystem
-
-object $name;format="space,Camel"$ContainerCmdApp extends App {
-
-   ContainerCmd.start("$name;format="normalize"$-container-cmd-app", Subsystem.withNameInsensitive("$subsystem$"),args)
-
-}
