--- src/main/g8/$deploy_module$/src/main/scala/org/tmt/$subsystem__lower$/$deploy_module__word$/$name__space,Camel$HostConfigApp.scala
@@ -1,10 +1,0 @@
-package org.tmt.$subsystem;format="lower"$.$name;format="space,norm,word"$deploy;
-
-import csw.framework.deploy.hostconfig.HostConfig
-import csw.prefix.models.Subsystem
-
-object $name;format="space,Camel"$HostConfigApp extends App {
-
-   HostConfig.start("$name;format="normalize"$-host-config-app",Subsystem.withNameInsensitive("$subsystem$"), args)
-
-}
