+++ src/main/g8/$hcdComponentName__space,norm$/src/main/resources/reference.conf
