+++ src/main/g8/$hcdComponentName__space,norm$/src/test/resources/$hcdComponentName__space,Camel$Standalone.conf
@@ -1,0 +1,4 @@
+prefix = "$subsystem$.$hcdComponentName;format="package"$"
+componentType = hcd
+behaviorFactoryClassName = "$organization$.$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$;.$hcdComponentName;format="space,Camel"$BehaviorFactory"
+locationServiceUsage = RegisterOnly
