+++ src/main/g8/$hcdComponentName__space,norm$/src/test/resources/application.conf
