--- src/main/g8/$hcd_module$/src/main/scala/org/tmt/$subsystem__lower$/$hcd_module__word$/$hcdComponentName__space,Camel$BehaviorFactory.scala
@@ -1,13 +1,0 @@
-package org.tmt.$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$;
-
-import akka.actor.typed.scaladsl.ActorContext
-import csw.framework.models.CswContext
-import csw.framework.scaladsl.{ComponentBehaviorFactory, ComponentHandlers}
-import csw.command.client.messages.TopLevelActorMessage
-
-class $hcdComponentName;format="space,Camel"$BehaviorFactory extends ComponentBehaviorFactory {
-
-  override def handlers(ctx: ActorContext[TopLevelActorMessage], cswCtx: CswContext): ComponentHandlers =
-    new $hcdComponentName;format="space,Camel"$Handlers(ctx, cswCtx)
-
-}
