+++ src/main/g8/$name__space,norm$-deploy/src/main/resources/$name__space,Camel$HostConfig.conf
