+++ src/main/g8/$name__space,norm$-deploy/src/main/resources/application.conf
