+++ src/main/g8/$name__space,norm$-deploy/src/main/scala/$organization$/$subsystem__lower$/$name__space,norm$deploy/$name__space,Camel$ContainerCmdApp.scala
@@ -1,0 +1,10 @@
+package $organization$.$subsystem;format="lower"$.$name;format="space,norm,word"$deploy;
+
+import csw.framework.deploy.containercmd.ContainerCmd
+import csw.prefix.models.Subsystem
+
+object $name;format="space,Camel"$ContainerCmdApp extends App {
+
+   ContainerCmd.start("$name;format="normalize"$-container-cmd-app", Subsystem.withNameInsensitive("$subsystem$"),args)
+
+}
