+++ src/main/g8/default.properties
@@ -4,10 +4,4 @@ assemblyComponentName=$name;format="package"$Assembly
 hcdComponentName=$name;format="package"$Hcd
 organization=org.tmt
 subsystem=CSW
-version=0.0.1
-package=$organization;format="lower"$.$subsystem;format="lower"$
-deploy_module=$name;format="space,norm"$-deploy
-assembly_module=$assemblyComponentName;format="space,norm"$
-hcd_module=$hcdComponentName;format="space,norm"$
-scala_version=2.13.3
 verbatim=coursier csw-services.sh
